agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
